requirements.txt
pillow-heif
opencv-python
distance
imagehash
pillow